backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & ML
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.5
qdrant-client==1.12.1
tiktoken==0.5.2

# Data Processing
beautifulsoup4==4.12.3
lxml==5.1.0
Pillow==10.2.0
python-multipart==0.0.6

# Redis
redis==5.0.1
hiredis==2.3.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
